+++ iftttie/channels/base.py
@@ -1,16 +1,13 @@
 from __future__ import annotations
 
 from abc import ABCMeta, abstractmethod
-from typing import Any
-
-from iftttie import web
 
 
-# TODO: rename to service?
 class BaseChannel(metaclass=ABCMeta):
+    @property
     @abstractmethod
-    async def run(self, context: web.Context, **kwargs: Any):
+    async def events(self):
         """
-        Runs the channel. May raise unhandled exceptions.
+        Gets an asynchronous generator of events.
         """
         raise NotImplementedError()