+++ iftttie/runner.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from asyncio.tasks import gather, sleep
 from concurrent.futures import CancelledError
-from contextlib import suppress
 
 from aiohttp import ClientConnectorError
 from loguru import logger
@@ -17,8 +16,7 @@ async def run_channels(context: Context):
     """
 
     logger.info('Running channels…')
-    with suppress(CancelledError):
-        await gather(*[run_channel(context, channel) for channel in context.channels])
+    await gather(*[run_channel(context, channel) for channel in context.channels])
 
 
 async def run_channel(context: Context, channel: BaseChannel):
@@ -28,14 +26,17 @@ async def run_channel(context: Context, channel: BaseChannel):
     while True:
         logger.info('Running {channel}…', channel=channel)
         try:
-            await channel.run(context)
+            async for event in channel.events:
+                await context.trigger_event(event)
         except CancelledError:
             logger.info('Stopped channel {channel}.', channel=channel)
             break
-        except ClientConnectorError as e:
+        except (ConnectionError, ClientConnectorError) as e:
             logger.error('{channel} has raised a connection error:', channel=channel)
             logger.error('{e}', e=e)
         except Exception as e:
             logger.opt(exception=e).error('{channel} has failed.', channel=channel)
+        else:
+            pass  # TODO: reset error counter.
         logger.error('Restarting in a minute.')
         await sleep(60.0)  # TODO: something smarter.